Create a new Blob model.
This commit introduces a new Unit type called Blob, and converts the
Docker v2 sync code to use it instead of overloading the Image model
that is also used by v1 sync and publish. This will simplify the
concepts in Pulp and make the v2 distributor easier to write.
Conveniently, it also separates the locations of the Units on the
filesystem as well.

https://pulp.plan.io/issues/967

re #967