upload_unit returns the unit key for what got uploaded
It is helpful to know the unit_key data for the manifest that was uploaded
in order to facilitate further processing.

Typically, in an automated workflow, a docker image upload is immediately
followed by an image tag, and the digest is needed in that case.

The return data is similar to what the association code returns.

closes #3232
https://pulp.plan.io/issues/3232

(cherry picked from commit 18e6c1fe83e9686e54178067b0c213067975e9ee)