Fix race in docker v2 publish
Previously, the last two steps of docker v2 publish were:

- Put symlink in place pointing at redirect file under master directory
- Create the redirect file

That means there's a period of time during publish where the redirect
file for the repo doesn't exist at the expected path.  If crane loads
metadata at this time, it will fail to load metadata for the repo.

This commit fixes the issue by moving the symlink rewrite to the last
step.

closes #3357
https://pulp.plan.io/issues/3357