Distribute tags according to requested media type
Return the artifacts associated with a tag with type failover. Try
manifest list first, then manifest schema 2. If both fail, log the
warning and return a 404. This provides a convinient location for an
on-the-fly schema converter.

These changes are necessary to use the split-tag models for
distribution, so they will be merged as a part of #4172.

https://pulp.plan.io/issues/4172
re #4172