Problem: Publications are not needed for Docker plugin
Solution: Add 'repository_version' field to DockerDistribution

This patch allows users to create DockerDistributions that serve repository versions. A distribution can
be created with either a repository or repository version defined. If a repository is defined, the latest
version of that repository is always available at the distribution. If a repository version is specified,
then that repository version is always available at that distribution.

This patch updates an existing test and adds a new one. The two tests assert that a docker or podman pull
can be performed when a repository is used in a distribution and when a repository version is used.

This patch also fixes the tests to work with docker as the client.

closes: #4669
https://pulp.plan.io/issues/4669