Problem: docker plugin does not support lazy sync
Solution: add support for lazy sync

This patch adds ability to create remotes with policy='on_demand'. When such a remote is used to
sync a repository, the Blobs are not downloaded. Only manifests and manifest lists are saved at
sync time. Blobs are retrieved on demand from clients.

This patch also includes a set of tests that check that you can pull content from a repo that
was synced using a remote that has policy set to 'on_demand'. The tests assert that performing
a docker/podman pull results in more artifacts being stored in Pulp than before the image was
requested.

closes: #4174
https://pulp.plan.io/issues/4174