Flatten queries for content unit removal
https://pulp.plan.io/issues/4549

Replace the recursive pattern with a fixed number of larger queries.

Additionally, reorder the content removal to "top down". This will fail
more gracefully; failure leaves orphans (safe) rather than user-facing
unlinked content (unsafe). This requires the additional plugin step of
removing the explicitly given content, which is normally handled by pulp
platform.

A side effect of this change is the correction of a bug that did not
remove shared content, even if all linked content is removed.
https://pulp.plan.io/issues/5161

fixes #5161
fixes #4549