Use Downloader.extra_headers instead of session.headers
In recent change in nectar session in downloader was removed from
shared space for all threads and for every request new Session
object is created. To be able to provide same functionality
extra_headers dictionary was added to Downloader object. This commit
modifies code to use extra_headers instead of session.headers